Version 1.1
Updates in version 1.1 include using Twilio's PHP library for sending
txt messages to make it easier to send updates to multiple phone
numbers. It also reduces the reuse of code for connecting to the
database and allows updates to be sent when transiting between "Road
Closed" and "Travel Not Recommended" statuses.